Refactor XML backend components: reorganize imports, introduce new classes, and enhance data handling for improved structure and clarity
