+++ spinta/datasets/backends/dataframe/backends/xml/components/__init__.py
@@ -1,0 +1,15 @@
+from .row import Row, RowList, RowModelRef
+from .dask_xml import DaskXml
+from .row_formatter import RowFormatter
+from .row_meta_item import RowMetaItem
+from .xml import Xml
+
+__all__ = [
+    "DaskXml",
+    "Row",
+    "RowList",
+    "RowModelRef",
+    "RowMetaItem",
+    "RowFormatter",
+    "Xml",
+]
