+++ spinta/datasets/backends/dataframe/backends/xml/components/dask_xml.py
@@ -4,8 +4,8 @@ from typing import Any, Dict, Iterable, Mapping, Optional
 import dask.dataframe as dd
 
 from spinta.adapters.loaders import DataAdapter, DataAdapterError
-from spinta.datasets.backends.dataframe.backends.xml.adapter.row import RowList
-from spinta.datasets.backends.dataframe.backends.xml.domain.model import ModelHeader
+from spinta.datasets.backends.dataframe.backends.xml.components.row import RowList
+from spinta.datasets.backends.dataframe.backends.xml.model import ModelHeader
 from spinta.utils.schema import NA
