--- spinta/datasets/backends/dataframe/backends/xml/domain/__init__.py
@@ -1,10 +1,0 @@
-from .model import DataModel, ModelHeader, MetaItem, ModelRef, TransformationModel, ModelItem
-
-__all__ = [
-    "DataModel",
-    "MetaItem",
-    "ModelHeader",
-    "ModelRef",
-    "TransformationModel",
-    "ModelItem",
-]
